feat: Zone commands and example zone

+++ plugins/kubectl/env-destroy.sh
@@ -5,16 +5,18 @@
 : "${taito_env:?}"
 : "${taito_namespace:?}"
 
-"${taito_plugin_path}/util/use-context.sh" && \
+if "${taito_cli_path}/util/confirm-execution.sh" "kubectl-delete" "${name}"; then
+  "${taito_plugin_path}/util/use-context.sh" && \
 
-echo "Deleting secrets from Kubernetes" && \
-"${taito_plugin_path}/util/delete-secrets.sh" && \
+  echo "Deleting secrets from Kubernetes" && \
+  "${taito_plugin_path}/util/delete-secrets.sh" && \
 
-echo "Delete namespace ${taito_namespace} (Y/n)?" && \
-echo "WARNING: Do not delete the namespace if it contains also some other apps" && \
-read -r confirm && \
-if [[ "${confirm}" =~ ^[Yy]*$ ]]; then
-  (${taito_setv:?}; kubectl delete namespace "${taito_namespace}")
+  echo "Delete namespace ${taito_namespace} (Y/n)?" && \
+  echo "WARNING: Do not delete the namespace if it contains also some other apps" && \
+  read -r confirm && \
+  if [[ "${confirm}" =~ ^[Yy]*$ ]]; then
+    (${taito_setv:?}; kubectl delete namespace "${taito_namespace}")
+  fi
 fi && \
 
 # Call next command on command chain
